+++ SiH_ModLoader/Data/HideWearKind.cs
@@ -1,0 +1,9 @@
+﻿namespace SiH_ModLoader.Data
+{
+    public enum HideWearKind
+    {
+        Unknown = 0,
+        Tops = 1,
+        Bottom = 2,
+    }
+}
