+++ SiH_ModLoader/Data/ItemKind.cs
@@ -1,0 +1,16 @@
+﻿namespace SiH_ModLoader.Data
+{
+    public enum ItemKind
+    {
+        Unknown = 0,
+        CT = 1, //Clothes Top
+        CB = 2, //Clothes Bottom
+        CF = 3, //Clothes Full (one-piece)
+        UT = 10, //Underwear Top
+        UB = 11, //Underwear Bottom
+        UF = 12, //Underwear Full (one-piece)
+        SX = 20, //Socks
+        FT = 21, //Footwear
+        GV = 22, //Gloves
+    }
+}
